Adds Player#ship as a #ships index helper method
Adds Player#ship method as a helper method to access a specific ship
inside the array of a player's ships returned by Player#ships.
It makes heavy use of Ship.index for validation and translation of
ship identifiers(String, Symbol and Fixnum) to return the specified
ship passed as an argument.

Signed-off-by: João Soares <jsoaresgeral@gmail.com>