Add Ship #position and #placed? initial behaviour
Adds #position and #placed attributes providing support for checking
if a ship is already placed on the board and, if true, where.
All ships are created with position = nil and placed = false.
Adds a helper method #placed? that return the #placed attribute status.
Still lacks the implementation of Player#place to specify further
behaviour.

Signed-off-by: João Soares <jsoaresgeral@gmail.com>