Adds new scenarios to place ships feature
Adds the ship out of board scenario where the ship may be placed with a
good anchor position but the orientation choosen may throw the rest of
the ship to an out of board sector.
Adds the ship overlapping another ship scenario where the ship may be
positioned on a sector that already contains a ship and so overlapping
it.

Signed-off-by: João Soares <jsoaresgeral@gmail.com>