Removed Player#place support for position arrays
Player#place no longer supports passing a position array, this was an
undocumented feature and now it has been removed.
The return value of Grid#parse_pos and Grid#format_pos is still valid
but it now needs a splat operator in order to flat the array into its
propper arguments.

Signed-off-by: João Soares <jsoaresgeral@gmail.com>