Improves Row and Col to raise exceptions
Row.is_valid? and Col.is_valid? are new methods for argument validation
checking and delegate conversion functionality to their respective
already existent Row.to_idx and Col.to_idx methods which no longer are
responsible for it.
Changes Row and Col #initialize methods to use the above two methods to
convert and validate the argument used to create a new instance and
raise an error if not valid.

Signed-off-by: João Soares <jsoaresgeral@gmail.com>