Adds support to check ship position valid status
Adds Pos#valid? which uses Row#valid? if orientation is vertical and
Col#valid? if it is horizontal. Each method accepts a Fixnum size or a
ship and in the last case uses its size or length to detect if the
position is valid for that ship.
Adds Ship#length as an alias for Ship#size in order to use duck typing
when detecting wether a size of a ship was passed, since a Fixnum
responds to :size.

Signed-off-by: João Soares <jsoaresgeral@gmail.com>